Restructured bot
